feat(cli): Add validation for new cli options. Move click definitions to cli.py
Resolves: #65
